requirements.txt
python-telegram-bot==20.7
requests
faker
pyotp
cryptography
flask
gunicorn